requirements.txt
PyYAML==6.0.1
Pillow==10.1.0